Update main.css
If the browser window is too small, the Undiscord window will be mostly offscreen and won't be visible. This fixes that by changing position: fixed to position: right. I also changed right: 10px to right: 44px to make things align better and look better aesthetically. So now, no matter how small the browser window is, the right edge of the Undiscord window will always be 44px from the right edge of the browser. Before, when it was fixed, almost all of the Undiscord window was not visible and we could only see the left side of the Undiscord window. Again, this fix is for when the browser window is made small and the Undiscord window becomes hidden and off screen.