Update deleteDiscordMessages.user.js
A lot of people have been experiencing 401 errors for over a year. The reason for these 401 errors is that Undiscord is auto filling the authentication tokens input box with the person's Discord password. That is insane. Our Discord passwords should be inserted in Undiscord like that. The fix is extremely simple. Replace the word "password" with "text" and Undiscord will now work without 401 errors and also the user can feel better about not having their Discord password auto inserted in text boxes.