Update deleteDiscordMessages.user.js
For over a year, people have been experiencing 401 errors because Undiscord is auto filling the authentication token text box with the user's Discord password. This is very alarming because that is a security flaw. The fix is very simple. Replace the word "password" with the word "text". Now, Undiscord will work without any 401 errors and the user can feel safer knowing that their Discord password isn't being randomly inserted somewhere.

In addition to that, people supposedly had an issue where Undiscord would stop even though there were messages remaining that needed to be deleted.

Someone's proposed fix was to add:
const messagesRemaining = this.state.grandTotal - this.state.offset;

and then add a log.verb message to go along with it. Seems like a reasonable small, simple, and useful fix.